[US_IX] Small change to tasks home visits (Recidiviz/recidiviz-data#41692)
## Description of the change

A few months ago, we implemented a change to make tasks more accurate to
the policy by requiring a home visit within 30 days of an address
change. Since then though, we have received a few tickets, like [Recidiviz/recidiviz-data#8028]
(https://github.com/Recidiviz/recidiviz-dashboards/issues/8028), where
home visits are being surfaced even though the officer visited the
client at their new residence.

In order to prevent this from happening in the future, I am implementing
a two day grace period, so if there is an address change but the officer
visited the client's home in the last two days then we do not require a
home visit.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 09d513190e2e1dae1d3e0c51e938297f51b07e9e