[Source Tables] Enumerate / clarify all updates that will be applied to each source table (Recidiviz/recidiviz-data#41728)
## Description of the change

Makes a handful of changes so that the source table update "dry run"
will return all the info we need to do the actual update, allowing us
(in a subsequent change), to unify some of our code that identifies
source table changes.

This PR:
* Renames `SourceTableUpdateDryRunResult` to
`SourceTableWithRequiredUpdateTypes`
* Updates `SourceTableWithRequiredUpdateTypes` to capture ALL the
different update types that will be applied to a table, not just the
first / most important one we find.
* Eliminates `SourceTableUpdateType.NO_CHANGES` and
`SourceTableUpdateType.UPDATE_SCHEMA_WITH_CHANGES` which both existed to
express when we either had no changes or multiple changes (we can now
infer this when there are empty sets of update types or multiple listed
update types).
* Moves the core functionality of `_build_table_update_results_str` into
a helper on `SourceTableWithRequiredUpdateTypes` and restructures in a
way that fully enumerates all the found updates. Adds lots of tests for
this new function.
* Renames `dry_run()` on the source table updater to
`get_changes_to_apply_to_source_tables()`. This function will eventually
be also used in the `update_tables()` call, so it's not just for dry
runs.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prep work for Recidiviz/recidiviz-data#41360

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9629bfbe71cdcf1a75c4d7ecf0ad6a3149aeb0d7