[US_TX][Tasks] Refactors `contact_compliance_builder` to switch directionality and use pre-materialized view of contact cadence spans (Recidiviz/recidiviz-data#41710)
## Description of the change

Does a couple little refactors of `contact_compliance_builder` to set us
up to be able to build out the special critical understaffing policy for
quarterly contacts (see
https://github.com/Recidiviz/recidiviz-data/pull/41718):
1. **Switches the direction of `meets_criteria` for the output of
`contact_compliance_builder`** so that true spans are periods of time
when a person is _not_ in compliance/needs a contact. Originally we had
talked about this wider refactor not being necessary at this stage, but
I realized it would actually be easier to refactor this shared function
(though I left some of the other criteria that don't use this function
as is) so that I could re-use `contact_compliance_builder` for the
critical understaffing policy without repeating a lot of logic.
2. Accordingly switches the naming and tasks record references for all
criteria that use `contact_compliance_builder` to properly interpret the
new meets_criteria values (i.e., from `meets` --> `needs`)
3. Pulls part of the contact_compliance_builder logic that generates
spans of time with contact cadences for each client into a **separate
analyst_data view**, since this (a) this is shared logic across many
criteria queries, and (b) we will need to directly reference this view
to identify clients having quarterly contact requirements. Also added a
new param for "custom_contact_cadence_spans" to
`contact_compliance_builder`, so that we have the option to input a set
of contact requirement spans that don't follow the default logic from
`us_tx_contact_cadence_spans` (this comes in handy for the critical
understaffing policy PR
https://github.com/Recidiviz/recidiviz-data/pull/41718)

Validated in sandbox that there are no changes to the number of records
surfaced as needing contacts.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances Recidiviz/recidiviz-data#41708

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 04b9736084b9ae4cff6230bae44d84c15eec1624