add feature_variant and inverse_feature_variant to the outlier_status+archive view (Recidiviz/recidiviz-data#41815)
## Description of the change

Adds the two columns so that we know which metrics had feature variants
configured on a given day. Eventually, we're going to need to know if an
officer was surfaceable in the product as an outlier on a given day, and
part of that definition is whether their direct supervisor had the
relevant fv / didn't have the relevant inverse fv. Since the feature
variants configured for a metric can change day to day, it'll be more
accurate to have this in an archive view. We only really need this for
the archive- in the backend we can just use the config as usual, so I'm
not planning on adding it to the DB schema.

I thought about exporting the config separately since this is putting
the same value on every metric for every officer for every time period,
which is more than we need, but having it here is going to be easier for
implementing https://github.com/Recidiviz/recidiviz-data/issues/40650 so
that we don't have to create a whole new archive view to join in. This
view is pretty denormalized as it is already, so it still follows
precedent.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/40650

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 714c8d8f57843d9b05a650f52158a08565f9a01c