[US_TN] Change zero-day period handling in custodial authority normalization (Recidiviz/recidiviz-data#41768)
## Description of the change

Recidiviz/recidiviz-data#40490 added normalization logic that overrides the custodial authority
for IPs with "-T" in the admission reason, using the custodial authority
from the most recent "-P" period. Part of this logic involved filtering
out zero-day "P" periods, which was intended to ensure that if a "-T"
period was preceded by two "-P" periods with the same admission date,
and one of them was a zero-day period, we would use the non-zero-day
period for the override. However, there are cases where the _only_ "-P"
periods preceding a "-T" period are zero-day, and in these cases we
still want to override the "-T" period's custodial authority.

This change stops filtering out zero-day periods from the set of periods
that can be used for the override, and instead adds some custom sorting
so that the list of preceding "-P" periods is sorted first by admission
date, then by length (both descending). This means that when we pick the
most recent "-P" period, if two periods are equally recent but one is
zero-day, we use the non-zero-day one (which is the same behavior as
before). On the other hand, if the most recent period is zero-day and
_doesn't_ share an admission date with any other periods, then its
custodial authority _will_ be used for the override.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 409948c3a7adc322b59e58344cc84f643a28e407