ERS Ineligible Offenses and not 6 months from release (Recidiviz/recidiviz-data#41844)
## Description of the change

Add ERS criteria:
> Have no current or prior felony convictions or deferred adjudication,
including juvenile convictions for the offense that includes the use of
a child in the commission of a crime, nor any offenses that were pled
down to a lesser degree but may have included an intent to commit sexual
assault, bodily harm, etc.

Also add criteria to not surface clients for ARS/ERS that will end
supervision within 6 months because officers agree it's not worth
submitting paperwork for them.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3d51a7b94c96f0814f76bdc8f71dc14f4a77476e