[TES] Fix zero day/overlapping spans issue (Recidiviz/recidiviz-data#41695)
## Description of the change

Small tweaks to the `is_past_completion_date_criteria_builder` helper to
fix the validation errors for TES criteria that popped this morning. I'm
still validating that this change removes all of the relevant errors,
but I think this is ready for review in the meantime.

## Related issues

Closes Recidiviz/recidiviz-data#41604

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 28fdf5ce322983e1358c3b369d32a84ed50e73fd