[Insights] move SUPERVISION_STATE_METRICS_VIEW_BUILDER out of INSIGHTS_VIEW_BUILDERS_TO_EXPORT (Recidiviz/recidiviz-data#41923)
## Description of the change

Today, @morden35 and I discovered that we've been returning 400s for the
`https://application-data-import-qqec6jbn6a-uc.a.run.app/import/insights/US_[all
states]/supervision_state_metrics.json` endpoint for quite a while. I
realized this view was being exported but there was no DB entity for it,
which was triggering the error. The view was originally removed in
https://github.com/Recidiviz/recidiviz-data/pull/33278, but then it was
added back in https://github.com/Recidiviz/recidiviz-data/pull/34662 so
it could be used in the prototype app, and we forgot to stop exporting
it.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b1396bee7fc03b890b1e108e5303efdf7e57f639