[TES] Fix date overflow in supervision_continuous_employment_for_6_months criteria (Recidiviz/recidiviz-data#42023)
## Description of the change

There are some bad dates in `state_employment_period` which are crashing
the `supervision_continuous_employment_for_6_months` criteria. Add a
bandaid so we filter out any of these bad dates before trying to add 6
months and create date overflow. Successfully loaded the criteria to a
sandbox with no state code filter.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://recidiviz.slack.com/archives/C04G37PRQ10/p1746483072326559

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 247cc8dde37fe99269bbaad783da4cdbf4b012b5