[Sentencing] Add offenses from charge_record to Case Insights (Recidiviz/recidiviz-data#41451)
## Description of the change

There are some charges (e.g. Drug-Trafficking in Methamphetamine or
Amphetamine (200 grams to Less Than 400 grams)) that are present in
charge_record (which pulls from normalized_state.state_charge) but are
not currently in case_insights_rates. This is because
case_insights_rates populates based on sentence_cohort, and if a charge
isn't associated with any actual sentence cohort start (e.g. it wasn't
the most severe charge, or it wasn't part of an initial sentence) then
it doesn't make it into sentence_cohort.

In order to make sure we have a more comprehensive list of charges, we
should populate case_insights_rates with rows corresponding to all
charges in charge_record.

This PR does this by adding charges from `charge_record` to the list of
unique offense names when enumerating all combinations of offenses and
other attributes (as in Recidiviz/recidiviz-data#35487).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40754

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4a220bc1dee5d1cd119bd9d68def692df61047a9