[US_AR] Use level_1_incarceration_location_name for locations in resident record (Recidiviz/recidiviz-data#40903)
## Description of the change

The resident record view uses the facility field from
most_recent_incarceration_population_span_metrics_materialized as the
`facility_id` by default. For most states, this is a readable name, but
for AR, it's an ID code that can be mapped to a readable name in
incarceration_location_ids_to_names. This results in facilities showing
up downstream identified only by their numeric IDs, not with their
actual names.

This change updates the _RESIDENT_RECORD_INCARCERATION_CTE in the
resident record view to use the level_1_incarceration_location_name for
`facility_id` in AR, while leaving `facility_id` unchanged for all other
states.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8648ec0516e5e76d080bb902750206d670682dd7